requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
captcha==0.3
certifi==2020.4.5.2
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.1
Django==3.1
django-ckeditor==5.4.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-jet==1.0.8
django-js-asset==1.0.0
django-redis==4.12.1
django-tinymce==2.8.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
easygui==0.98.1
Faker==4.1.1
feedparser==5.2.1
filelock==3.0.12
flake8==3.8.3
future==0.18.2
google-api-python-client==1.4.1
httplib2==0.18.1
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.6.1
isort==4.3.21
jwt==1.0.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
moviepy==1.0.3
mysqlclient==2.0.1
numpy==1.18.5
oauth2client==4.1.3
Pillow==7.1.2
proglog==0.1.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.4
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.46.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.27
wrapt==1.12.1
zipp==3.1.0
